feat: remove structure
