+++ lib/nonnative/process.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Nonnative
+  class Process
+    attr_accessor :command
+    attr_accessor :timeout
+    attr_accessor :port
+    attr_accessor :file
+  end
+end
